# Cherry-pick #20717 to 7.9: Improve registry file migration performance